Enhance getValue Docs (#329)
* Implement "Types"

* Remove unused lodash

* enhance getValues docs
