+++ src/components/codeExamples/getValues.ts
@@ -12,9 +12,10 @@ export default function App() {
       <button
         type="button"
         onClick={() => {
-          const values = getValues();
-          const singleValue = getValues("test");
-          const singleValue = getValues(["test", "test1"]);
+          const values = getValues(); // { test: "test-input", test1: "test1-input" }
+          const singleValue = getValues("test"); // "test-input"
+          const singleValue = getValues(["test", "test1"]); 
+          // { test: "test-input", test1: "test1-input" }
         }}
       >
         Get Values
